Move duplicate BlockList and TxList to List.tsx
